Use oracle files to test ungroup sync
