--- packages/haiku-serialization/test/fixtures/projects/tgroup/code/main/lottie.json
@@ -1,1 +1,0 @@
-{"ip":0,"fr":60,"v":"5.1.20","assets":[],"layers":[{"ty":4,"nm":"Rectangle B","ip":0,"st":0,"ind":9,"hix":1,"ks":{"o":{"a":0,"k":100},"or":{"a":0,"k":[0,0,0]},"a":{"a":0,"k":[123,48.5,0]},"p":{"s":true,"x":{"a":0,"k":427.1630859375},"y":{"a":0,"k":373.5675}},"rx":{"a":0,"k":0},"ry":{"a":0,"k":0},"rz":{"a":0,"k":0},"s":{"a":0,"k":[58.3,54.50000000000001]}},"shapes":[{"ty":"gr","it":[{"ty":"rc","s":{"a":0,"k":[246,97]},"r":{"a":0,"k":0},"p":{"a":0,"k":[123,48.5]}},{"ty":"st","o":{"a":0,"k":0},"w":{"a":0,"k":0},"c":{"a":0,"k":[0,0,0,0]},"lc":3,"lj":1},{"ty":"fl","o":{"a":0,"k":100},"r":1,"c":{"a":0,"k":[0,0,1,1]}},{"ty":"tr","o":{"a":0,"k":100},"a":{"a":0,"k":[0,0]},"s":{"a":0,"k":[100,100]},"p":{"a":0,"k":[0,0]},"r":{"a":0,"k":0}}]}],"op":1},{"ty":4,"nm":"Group Inner","ip":0,"st":0,"ind":8,"hix":2,"ks":{"o":{"a":0,"k":100},"or":{"a":0,"k":[0,0,0]},"a":{"a":0,"k":[236,93.5,0]},"p":{"s":true,"x":{"a":0,"k":412.76402539062497},"y":{"a":0,"k":288.2904968261719}},"rx":{"a":0,"k":0},"ry":{"a":0,"k":0},"rz":{"a":0,"k":0},"s":{"a":0,"k":[58.3,54.50000000000001]}},"shapes":[{"ty":"gr","it":[{"ty":"rc","s":{"a":0,"k":[276,187]},"r":{"a":0,"k":43},"p":{"a":0,"k":[138,93.5]}},{"ty":"st","o":{"a":0,"k":0},"w":{"a":0,"k":0},"c":{"a":0,"k":[0,0,0,0]},"lc":3,"lj":1},{"ty":"fl","o":{"a":0,"k":100},"r":1,"c":{"a":0,"k":[0,1,0,1]}},{"ty":"tr","o":{"a":0,"k":100},"a":{"a":0,"k":[0,0]},"s":{"a":0,"k":[100,100]},"p":{"a":0,"k":[0,0]},"r":{"a":0,"k":0}}]},{"ty":"gr","it":[{"ty":"rc","s":{"a":0,"k":[78,59]},"r":{"a":0,"k":0},"p":{"a":0,"k":[39,29.5]}},{"ty":"st","o":{"a":0,"k":0},"w":{"a":0,"k":0},"c":{"a":0,"k":[0,0,0,0]},"lc":3,"lj":1},{"ty":"fl","o":{"a":0,"k":100},"r":1,"c":{"a":0,"k":[1,0,0,1]}},{"ty":"tr","o":{"a":0,"k":100},"a":{"a":0,"k":[0,0]},"s":{"a":0,"k":[100,100]},"p":{"a":0,"k":[394,103]},"r":{"a":0,"k":0}}]}],"op":1},{"ty":4,"nm":"Rectangle G","ip":0,"st":0,"ind":7,"hix":3,"ks":{"o":{"a":0,"k":100},"or":{"a":0,"k":[0,0,0]},"a":{"a":0,"k":[138,93.5,0]},"p":{"s":true,"x":{"a":0,"k":80.454},"y":{"a":0,"k":250.958}},"rx":{"a":0,"k":0},"ry":{"a":0,"k":0},"rz":{"a":0,"k":0},"s":{"a":0,"k":[58.3,54.50000000000001]}},"shapes":[{"ty":"gr","it":[{"ty":"rc","s":{"a":0,"k":[276,187]},"r":{"a":0,"k":43},"p":{"a":0,"k":[138,93.5]}},{"ty":"st","o":{"a":0,"k":0},"w":{"a":0,"k":0},"c":{"a":0,"k":[0,0,0,0]},"lc":3,"lj":1},{"ty":"fl","o":{"a":0,"k":100},"r":1,"c":{"a":0,"k":[0,1,0,1]}},{"ty":"tr","o":{"a":0,"k":100},"a":{"a":0,"k":[0,0]},"s":{"a":0,"k":[100,100]},"p":{"a":0,"k":[0,0]},"r":{"a":0,"k":0}}]}],"op":1},{"ty":4,"nm":"Rectangle R","ip":0,"st":0,"ind":6,"hix":4,"ks":{"o":{"a":0,"k":100},"or":{"a":0,"k":[0,0,0]},"a":{"a":0,"k":[39,29.5,0]},"p":{"s":true,"x":{"a":0,"k":252.439},"y":{"a":0,"k":272.213}},"rx":{"a":0,"k":0},"ry":{"a":0,"k":0},"rz":{"a":0,"k":0},"s":{"a":0,"k":[58.3,54.50000000000001]}},"shapes":[{"ty":"gr","it":[{"ty":"rc","s":{"a":0,"k":[78,59]},"r":{"a":0,"k":0},"p":{"a":0,"k":[39,29.5]}},{"ty":"st","o":{"a":0,"k":0},"w":{"a":0,"k":0},"c":{"a":0,"k":[0,0,0,0]},"lc":3,"lj":1},{"ty":"fl","o":{"a":0,"k":100},"r":1,"c":{"a":0,"k":[1,0,0,1]}},{"ty":"tr","o":{"a":0,"k":100},"a":{"a":0,"k":[0,0]},"s":{"a":0,"k":[100,100]},"p":{"a":0,"k":[0,0]},"r":{"a":0,"k":0}}]}],"op":1},{"ty":4,"nm":"Rectangle B","ip":0,"st":0,"ind":5,"hix":5,"ks":{"o":{"a":0,"k":100},"or":{"a":0,"k":[0,0,0]},"a":{"a":0,"k":[123,48.5,0]},"p":{"s":true,"x":{"a":0,"k":152.1629794921875},"y":{"a":0,"k":373.5675}},"rx":{"a":0,"k":0},"ry":{"a":0,"k":0},"rz":{"a":0,"k":0},"s":{"a":0,"k":[58.3,54.50000000000001]}},"shapes":[{"ty":"gr","it":[{"ty":"rc","s":{"a":0,"k":[246,97]},"r":{"a":0,"k":0},"p":{"a":0,"k":[123,48.5]}},{"ty":"st","o":{"a":0,"k":0},"w":{"a":0,"k":0},"c":{"a":0,"k":[0,0,0,0]},"lc":3,"lj":1},{"ty":"fl","o":{"a":0,"k":100},"r":1,"c":{"a":0,"k":[0,0,1,1]}},{"ty":"tr","o":{"a":0,"k":100},"a":{"a":0,"k":[0,0]},"s":{"a":0,"k":[100,100]},"p":{"a":0,"k":[0,0]},"r":{"a":0,"k":0}}]}],"op":1},{"ty":4,"nm":"Rectangle B","ip":0,"st":0,"ind":4,"hix":6,"ks":{"o":{"a":0,"k":100},"or":{"a":0,"k":[0,0,0]},"a":{"a":0,"k":[123,48.5,0]},"p":{"s":true,"x":{"a":0,"k":427.163},"y":{"a":0,"k":173.568}},"rx":{"a":0,"k":0},"ry":{"a":0,"k":0},"rz":{"a":0,"k":0},"s":{"a":0,"k":[58.3,54.50000000000001]}},"shapes":[{"ty":"gr","it":[{"ty":"rc","s":{"a":0,"k":[246,97]},"r":{"a":0,"k":0},"p":{"a":0,"k":[123,48.5]}},{"ty":"st","o":{"a":0,"k":0},"w":{"a":0,"k":0},"c":{"a":0,"k":[0,0,0,0]},"lc":3,"lj":1},{"ty":"fl","o":{"a":0,"k":100},"r":1,"c":{"a":0,"k":[0,0,1,1]}},{"ty":"tr","o":{"a":0,"k":100},"a":{"a":0,"k":[0,0]},"s":{"a":0,"k":[100,100]},"p":{"a":0,"k":[0,0]},"r":{"a":0,"k":0}}]}],"op":1},{"ty":4,"nm":"Group Inner","ip":0,"st":0,"ind":3,"hix":7,"ks":{"o":{"a":0,"k":100},"or":{"a":0,"k":[0,0,0]},"a":{"a":0,"k":[236,93.5,0]},"p":{"s":true,"x":{"a":0,"k":412.588},"y":{"a":0,"k":50.943}},"rx":{"a":0,"k":0},"ry":{"a":0,"k":0},"rz":{"a":0,"k":0},"s":{"a":0,"k":[58.3,54.50000000000001]}},"shapes":[{"ty":"gr","it":[{"ty":"rc","s":{"a":0,"k":[276,187]},"r":{"a":0,"k":43},"p":{"a":0,"k":[138,93.5]}},{"ty":"st","o":{"a":0,"k":0},"w":{"a":0,"k":0},"c":{"a":0,"k":[0,0,0,0]},"lc":3,"lj":1},{"ty":"fl","o":{"a":0,"k":100},"r":1,"c":{"a":0,"k":[0,1,0,1]}},{"ty":"tr","o":{"a":0,"k":100},"a":{"a":0,"k":[0,0]},"s":{"a":0,"k":[100,100]},"p":{"a":0,"k":[0,0]},"r":{"a":0,"k":0}}]},{"ty":"gr","it":[{"ty":"rc","s":{"a":0,"k":[78,59]},"r":{"a":0,"k":0},"p":{"a":0,"k":[39,29.5]}},{"ty":"st","o":{"a":0,"k":0},"w":{"a":0,"k":0},"c":{"a":0,"k":[0,0,0,0]},"lc":3,"lj":1},{"ty":"fl","o":{"a":0,"k":100},"r":1,"c":{"a":0,"k":[1,0,0,1]}},{"ty":"tr","o":{"a":0,"k":100},"a":{"a":0,"k":[0,0]},"s":{"a":0,"k":[100,100]},"p":{"a":0,"k":[394,103]},"r":{"a":0,"k":0}}]}],"op":1},{"ty":4,"nm":"Group Outer","ip":0,"st":0,"ind":1,"hix":8,"ks":{"o":{"a":0,"k":100},"or":{"a":0,"k":[0,0,0]},"a":{"a":0,"k":[236,183.5,0]},"p":{"s":true,"x":{"a":0,"k":137.4120030179253},"y":{"a":0,"k":99.99299654267134}},"rx":{"a":0,"k":0},"ry":{"a":0,"k":0},"rz":{"a":0,"k":0},"s":{"a":0,"k":[58.3,54.50000000000001]}},"shapes":[{"ty":"gr","it":[{"ty":"rc","s":{"a":0,"k":[246,97]},"r":{"a":0,"k":0},"p":{"a":0,"k":[123,48.5]}},{"ty":"st","o":{"a":0,"k":0},"w":{"a":0,"k":0},"c":{"a":0,"k":[0,0,0,0]},"lc":3,"lj":1},{"ty":"fl","o":{"a":0,"k":100},"r":1,"c":{"a":0,"k":[0,0,1,1]}},{"ty":"tr","o":{"a":0,"k":100},"a":{"a":0,"k":[0,0]},"s":{"a":0,"k":[100,100]},"p":{"a":0,"k":[138,270]},"r":{"a":0,"k":0}}]},{"ty":"gr","it":[{"ty":"rc","s":{"a":0,"k":[276,187]},"r":{"a":0,"k":43},"p":{"a":0,"k":[138,93.5]}},{"ty":"st","o":{"a":0,"k":0},"w":{"a":0,"k":0},"c":{"a":0,"k":[0,0,0,0]},"lc":3,"lj":1},{"ty":"fl","o":{"a":0,"k":100},"r":1,"c":{"a":0,"k":[0,1,0,1]}},{"ty":"tr","o":{"a":0,"k":100},"a":{"a":0,"k":[0,0]},"s":{"a":0,"k":[100,100]},"p":{"a":0,"k":[0,0]},"r":{"a":0,"k":0}}]},{"ty":"gr","it":[{"ty":"rc","s":{"a":0,"k":[78,59]},"r":{"a":0,"k":0},"p":{"a":0,"k":[39,29.5]}},{"ty":"st","o":{"a":0,"k":0},"w":{"a":0,"k":0},"c":{"a":0,"k":[0,0,0,0]},"lc":3,"lj":1},{"ty":"fl","o":{"a":0,"k":100},"r":1,"c":{"a":0,"k":[1,0,0,1]}},{"ty":"tr","o":{"a":0,"k":100},"a":{"a":0,"k":[0,0]},"s":{"a":0,"k":[100,100]},"p":{"a":0,"k":[394,103]},"r":{"a":0,"k":0}}]}],"op":1}],"op":1,"w":550,"h":400}
