+++ packages/haiku-serialization/test/fixtures/projects/tgroup/designs/ciK4deII3q0dSMfJDUxoHaW2-t1.figma.contents/groups/Group Outer.svg.after-sync.1
@@ -1,0 +1,9 @@
+<svg width="472" height="367" viewBox="0 0 472 367" fill="none" xmlns="http://www.w3.org/2000/svg">
+<g id="Group Outer">
+<g id="Group Inner">
+<rect id="Rectangle R" width="100" height="59" rx="11" transform="translate(333 103)" fill="#FF0000"/>
+<rect id="Rectangle G" width="276" height="187" fill="#00FF00"/>
+</g>
+<rect id="Rectangle B" width="200" height="97" rx="22" transform="translate(138 222)" fill="#0000FF"/>
+</g>
+</svg>
